requirements.txt
fastapi
motor
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
uvicorn
pydantic[email]